Fix a grand process leak in examples rendering
During writing of cache heater excessive memory usage
became visible. One of the main reason was creation (and never destroying)
a new coffeescript process for highlighting (which of course created
a lot of processes eating RAM as we have a lot of examples).
This commit takes special measures to use just a single instance
of the highlighter process up.