Streamline CSS minifying
Previously `css-release` executed css-build and css-optimize.
The css-optimize command due to a filter did not include main.css,
which created a lot of confusion. Not only that, but it replaced
the content of the files built with minimified version instead
of creating minimified files alongside.

To streamline this situation here we do the following:

* If the dev wants to see non-mini css, doing `css-build` command is advised
* Normally minified css is created out of scss which is commited to the repo,
  including both main.css and other files, *.min.css files are removed