Use host network mode for docker setup
To workaroud issue with running examples, we provide our own
small server to dispatch the requests to executing containers.
To do this we need host network, so switch our containers
to this mode.

Other than this, use a pre-built image for the website
to reduce time of deploying updates, as a more powerful machine
can build an image to download on the actual server.