Hide sidebar on start if the window is too small
If the sidebar position is not set by the user before
and we are just opening the page for the first time,
check if we have enough width for it not to overlap with the content.

Along the way, fix warning related to setting incorrect value for cookies SameSite param.