Add missing timeout for search input unfocus
Prior to that, clicking on the "search in google" 'a' element
led to the whole search input to receive unfocus event and
immediately begin to hide the results div - so the 'a' element
was already hidden when it was its turn to process the click
event.

Here we add a small timeout for the unfocus so that the 'a' element's click
event had some time to be processed (so opening a new page).

It is indeed a hack and adds a small visual glitch, but no better solution for now
is present.